Moving GithubRepo to use yaml config. Working on #78
