+++ spec/DevBoard/Github/Repo/Entity/GithubRepoRepositorySpec.php
@@ -1,5 +1,5 @@
 <?php
-namespace spec\DevBoard\Github\Repo;
+namespace spec\DevBoard\Github\Repo\Entity;
 
 use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\Mapping\ClassMetadata;
@@ -10,7 +10,7 @@ class GithubRepoRepositorySpec extends ObjectBehavior
 {
     public function it_is_initializable()
     {
-        $this->shouldHaveType('DevBoard\Github\Repo\GithubRepoRepository');
+        $this->shouldHaveType('DevBoard\Github\Repo\Entity\GithubRepoRepository');
     }
 
     public function let(EntityManager $em, ClassMetadata $classMetadata)
