+++ src/DevBoard/Behat/Github/Repo/FixtureContext.php
@@ -2,7 +2,7 @@
 namespace DevBoard\Behat\Github\Repo;
 
 use Behat\Gherkin\Node\TableNode;
-use DevBoard\Github\Repo\GithubRepo;
+use DevBoard\Github\Repo\Entity\GithubRepo;
 use Resources\Behat\DefaultContext;
 
 /**
