+++ src/DevBoard/Core/Project/Entity/Project.php
@@ -1,7 +1,7 @@
 <?php
 namespace DevBoard\Core\Project\Entity;
 
-use DevBoard\Github\Repo\GithubRepo;
+use DevBoard\Github\Repo\Entity\GithubRepo;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\Mapping as ORM;
 use NullDev\UserBundle\Entity\User;
