+++ src/DevBoard/Github/Branch/Entity/GithubBranch.php
@@ -2,7 +2,7 @@
 namespace DevBoard\Github\Branch\Entity;
 
 use DevBoard\Github\Commit\GithubCommit;
-use DevBoard\Github\Repo\GithubRepo;
+use DevBoard\Github\Repo\Entity\GithubRepo;
 use Doctrine\ORM\Mapping as ORM;
 use Resources\Entity\BaseEntity;
